@@ -12,7 +12,7 @@
 	"devDependencies": {
 		"@typescript-eslint/eslint-plugin": "^5.48.1",
 		"@typescript-eslint/parser": "^5.48.1",
-		"eslint": "^8.31.0",
+		"eslint": "^8.32.0",
 		"eslint-plugin-solid": "^0.9.2",
 		"typescript": "^4.9.4",
 		"vite": "^4.0.4",
@@ -21,6 +21,8 @@
 	},
 	"dependencies": {
 		"@solidjs/router": "^0.6.0",
+		"bootstrap": "^5.2.3",
+		"solid-bootstrap": "^1.0.13",
 		"solid-js": "^1.6.9"
 	}
 }
