@@ -1,7 +1,6 @@
 /* @refresh reload */
-import { render } from "solid-js/web";
-
 import { Router } from "@solidjs/router";
+import { render } from "solid-js/web";
 import App from "./App";
 
 render(
